framework/front_end/requirements.txt
# --- Core UI ---
streamlit>=1.28
streamlit-autorefresh>=0.0.2

# --- Imaging ---
pillow>=10.0.0

# --- HTTP client (likely used inside client.py) ---
requests>=2.31.0

# --- Optional ---
# python-dotenv>=1.0.0     # only if you read a .env locally

